feat: 构建分组并发时优化 #9618
